# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KanBlam API",
    "version": "1.0",
    "description": "Token-authenticated REST API over your KanBlam workspace. Create a personal access token in Settings → API tokens and send it as `Authorization: Bearer kb_…`. A token acts as its user, limited by its scopes (`read`, `write`). Rate limit: 120 requests/minute per token."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal access token from Settings → API tokens (kb_…)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "invalid_request",
                  "rate_limited",
                  "internal"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "List tasks (filterable, cursor-paginated)",
        "description": "Stable ordering (createdAt desc). Follow `nextCursor` until it is null.",
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assigneeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stageId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quadrant",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "q1",
                "q2",
                "q3",
                "q4"
              ]
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hideCompleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of tasks",
            "content": {
              "application/json": {
                "example": {
                  "tasks": [
                    {
                      "id": "cku…task",
                      "name": "Tune 2 m dipole antenna",
                      "description": null,
                      "notes": null,
                      "project": {
                        "id": "cku…proj",
                        "name": "Payload & Electronics",
                        "code": "PAY"
                      },
                      "stage": {
                        "id": "cku…stage",
                        "name": "In Progress",
                        "isTerminal": false
                      },
                      "priority": {
                        "id": "cku…prio",
                        "name": "Medium"
                      },
                      "assignee": {
                        "id": "cku…user",
                        "name": "Brisk Otter"
                      },
                      "isImportant": true,
                      "isUrgent": false,
                      "quadrant": "q2",
                      "startDate": null,
                      "dueDate": "2026-08-01",
                      "progressPct": 50,
                      "progressManual": false,
                      "recurring": false,
                      "tags": [
                        {
                          "id": "cku…tag",
                          "name": "ham-radio",
                          "color": "#8b5cf6"
                        }
                      ],
                      "subtasks": [
                        {
                          "id": "cku…sub",
                          "title": "Trim both legs",
                          "completed": true,
                          "position": 0
                        }
                      ],
                      "createdAt": "2026-07-10T18:37:13.553Z",
                      "updatedAt": "2026-07-10T18:37:13.553Z"
                    }
                  ],
                  "nextCursor": null
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, revoked, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — see Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create a task",
        "description": "Minimal body: `{projectId, name}`. Omitted priority → the workspace's Medium; omitted stage → first non-terminal stage; omitted assignee → the token's user.",
        "x-required-scope": "write",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 10000
                  },
                  "isImportant": {
                    "type": "boolean"
                  },
                  "isUrgent": {
                    "type": "boolean"
                  },
                  "priorityId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "kanbanStageId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "assigneeId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tagIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "startDate": {
                    "type": "string"
                  },
                  "dueDate": {
                    "type": "string"
                  },
                  "progressPct": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 50000
                  },
                  "subtasks": {
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        }
                      },
                      "required": [
                        "title"
                      ]
                    }
                  }
                },
                "required": [
                  "projectId",
                  "name",
                  "priorityId",
                  "kanbanStageId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "example": {
                  "task": {
                    "id": "cku…task",
                    "name": "Tune 2 m dipole antenna",
                    "description": null,
                    "notes": null,
                    "project": {
                      "id": "cku…proj",
                      "name": "Payload & Electronics",
                      "code": "PAY"
                    },
                    "stage": {
                      "id": "cku…stage",
                      "name": "In Progress",
                      "isTerminal": false
                    },
                    "priority": {
                      "id": "cku…prio",
                      "name": "Medium"
                    },
                    "assignee": {
                      "id": "cku…user",
                      "name": "Brisk Otter"
                    },
                    "isImportant": true,
                    "isUrgent": false,
                    "quadrant": "q2",
                    "startDate": null,
                    "dueDate": "2026-08-01",
                    "progressPct": 50,
                    "progressManual": false,
                    "recurring": false,
                    "tags": [
                      {
                        "id": "cku…tag",
                        "name": "ham-radio",
                        "color": "#8b5cf6"
                      }
                    ],
                    "subtasks": [
                      {
                        "id": "cku…sub",
                        "title": "Trim both legs",
                        "completed": true,
                        "position": 0
                      }
                    ],
                    "createdAt": "2026-07-10T18:37:13.553Z",
                    "updatedAt": "2026-07-10T18:37:13.553Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, revoked, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found (cross-workspace ids are indistinguishable from unknown ids)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — see Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
  }
}